+++ ui/Changelog.md
@@ -1,8 +1,9 @@
 # Changelog
 
 ## 0.3.1
-### New Feature:
+### New Features:
 - The add new node menu can now be filtered by typing the part of a name of the node
+- Can now insert a graph node into connections by dropping it on top of a compatible connection
 
 ### Improvements:
 - "Dissolve Node" in the node context menu now works, it removes the node and tries to connect it's inputs to where it's output was connected